Moved DateTime to its own container in components, and moved the appropriate CSS for it to its own .css file. DateTime can now be called to home with the container DateTime.
